+++ packages/jaspr_cli/lib/src/migrations/migration_models.dart
@@ -1,9 +1,10 @@
-import 'dart:io';
+
 
 import 'package:analyzer/dart/analysis/utilities.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/token.dart';
 import 'package:analyzer/source/line_info.dart';
+import 'package:file/file.dart';
 
 abstract class Migration {
   String get name;
@@ -12,18 +13,21 @@ abstract class Migration {
   String get hint;
 
   void runForUnit(CompilationUnit unit, MigrationReporter reporter);
+
+  List<MigrationResult> runForDirectory(Directory dir, bool apply);
 }
 
 extension MigrationExtension on List<Migration> {
   List<MigrationResult> computeResults(
     List<String> directories,
     bool apply,
+    FileSystem fs,
     void Function(File, Object, StackTrace) onError,
   ) {
     final results = <MigrationResult>[];
 
     for (final path in directories) {
-      final dir = Directory(path);
+      final dir = fs.directory(path);
       if (!dir.existsSync()) {
         continue;
       }
@@ -35,7 +39,7 @@ extension MigrationExtension on List<Migration> {
         try {
           final result = parseString(content: content);
 
-          final builder = MigrationBuilder(result.lineInfo);
+          final builder = EditBuilder(result.lineInfo);
           final reporter = MigrationReporter(builder);
 
           for (final migration in this) {
@@ -46,7 +50,7 @@ extension MigrationExtension on List<Migration> {
             continue;
           }
 
-          results.add(MigrationResult(file.path, content, reporter));
+          results.add(MigrationResult(file.path, reporter.migrations, reporter.warnings));
 
           if (apply) {
             final result = builder.apply(content);
@@ -56,6 +60,11 @@ extension MigrationExtension on List<Migration> {
           onError(file, e, st);
         }
       }
+
+      for (final migration in this) {
+        results.addAll(migration.runForDirectory(dir, apply));
+      }
+
     }
 
     return results;
@@ -63,17 +72,17 @@ extension MigrationExtension on List<Migration> {
 }
 
 class MigrationResult {
-  MigrationResult(this.path, this.content, this.reporter);
+  MigrationResult(this.path, this.migrations, this.warnings);
 
   final String path;
-  final String content;
-  final MigrationReporter reporter;
+  final List<MigrationInstance> migrations;
+  final List<MigrationWarning> warnings;
 }
 
 class MigrationReporter {
   MigrationReporter(this._builder);
 
-  final MigrationBuilder _builder;
+  final EditBuilder _builder;
   final List<MigrationInstance> migrations = [];
   final List<MigrationWarning> warnings = [];
 
@@ -88,7 +97,7 @@ class MigrationReporter {
     warnings.add(MigrationWarning(offset, length, message, _currentMigration!));
   }
 
-  void createMigration(String description, void Function(MigrationBuilder builder) migration) {
+  void createMigration(String description, void Function(EditBuilder builder) migration) {
     migrations.add(MigrationInstance(description, _currentMigration!));
     migration(_builder);
   }
@@ -113,8 +122,8 @@ class MigrationInstance {
   String toString() => description;
 }
 
-class MigrationBuilder {
-  MigrationBuilder(this.lineInfo);
+class EditBuilder {
+  EditBuilder(this.lineInfo);
 
   final LineInfo lineInfo;
   final List<ChangeRequest> changes = [];